+++ config/config.go
@@ -14,6 +14,7 @@
 package config
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"regexp"
@@ -33,22 +34,26 @@ func LoadFile(filename string) (*Config, error) {
 	if err != nil {
 		return nil, err
 	}
+	if cfg.Version != 2 {
+		return cfg, ErrUnsupportedVersion
+	}
 	return cfg, nil
 }
 
 var (
+	defaultRetries = 3
+
 	DefaultAuth = Auth{
 		Community:     "public",
 		SecurityLevel: "noAuthNoPriv",
 		AuthProtocol:  "MD5",
 		PrivProtocol:  "DES",
+		Version:       2,
 	}
 	DefaultWalkParams = WalkParams{
-		Version:                 2,
 		MaxRepetitions:          25,
-		Retries:                 3,
+		Retries:                 &defaultRetries,
 		Timeout:                 time.Second * 5,
-		Auth:                    DefaultAuth,
 		UseUnconnectedUDPSocket: false,
 		AllowNonIncreasingOIDs:  false,
 	}
@@ -58,72 +63,42 @@ var (
 	DefaultRegexpExtract = RegexpExtract{
 		Value: "$1",
 	}
+
+	ErrUnsupportedVersion = errors.New("unsupported config version")
 )
 
 // Config for the snmp_exporter.
-type Config map[string]*Module
+type Config struct {
+	Auths   map[string]*Auth   `yaml:"auths,omitempty"`
+	Modules map[string]*Module `yaml:"modules,omitempty"`
+	Version int                `yaml:"version"`
+}
 
 type WalkParams struct {
-	Version                 int           `yaml:"version,omitempty"`
 	MaxRepetitions          uint32        `yaml:"max_repetitions,omitempty"`
-	Retries                 int           `yaml:"retries,omitempty"`
+	Retries                 *int          `yaml:"retries,omitempty"`
 	Timeout                 time.Duration `yaml:"timeout,omitempty"`
-	Auth                    Auth          `yaml:"auth,omitempty"`
 	UseUnconnectedUDPSocket bool          `yaml:"use_unconnected_udp_socket,omitempty"`
 	AllowNonIncreasingOIDs  bool          `yaml:"allow_nonincreasing_oids,omitempty"`
 }
 
 type Module struct {
 	// A list of OIDs.
-	Walk       []string   `yaml:"walk,omitempty"`
-	Get        []string   `yaml:"get,omitempty"`
-	Metrics    []*Metric  `yaml:"metrics"`
-	WalkParams WalkParams `yaml:",inline"`
+	Walk       []string        `yaml:"walk,omitempty"`
+	Get        []string        `yaml:"get,omitempty"`
+	Metrics    []*Metric       `yaml:"metrics"`
+	WalkParams WalkParams      `yaml:",inline"`
+	Filters    []DynamicFilter `yaml:"filters,omitempty"`
 }
 
 func (c *Module) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	*c = DefaultModule
 	type plain Module
-	if err := unmarshal((*plain)(c)); err != nil {
-		return err
-	}
-
-	wp := c.WalkParams
-
-	if wp.Version < 1 || wp.Version > 3 {
-		return fmt.Errorf("SNMP version must be 1, 2 or 3. Got: %d", wp.Version)
-	}
-	if wp.Version == 3 {
-		switch wp.Auth.SecurityLevel {
-		case "authPriv":
-			if wp.Auth.PrivPassword == "" {
-				return fmt.Errorf("priv password is missing, required for SNMPv3 with priv")
-			}
-			if wp.Auth.PrivProtocol != "DES" && wp.Auth.PrivProtocol != "AES" && wp.Auth.PrivProtocol != "AES192" && wp.Auth.PrivProtocol != "AES192C" && wp.Auth.PrivProtocol != "AES256" && wp.Auth.PrivProtocol != "AES256C" {
-				return fmt.Errorf("priv protocol must be DES or AES")
-			}
-			fallthrough
-		case "authNoPriv":
-			if wp.Auth.Password == "" {
-				return fmt.Errorf("auth password is missing, required for SNMPv3 with auth")
-			}
-			if wp.Auth.AuthProtocol != "MD5" && wp.Auth.AuthProtocol != "SHA" && wp.Auth.AuthProtocol != "SHA224" && wp.Auth.AuthProtocol != "SHA256" && wp.Auth.AuthProtocol != "SHA384" && wp.Auth.AuthProtocol != "SHA512" {
-				return fmt.Errorf("auth protocol must be SHA or MD5")
-			}
-			fallthrough
-		case "noAuthNoPriv":
-			if wp.Auth.Username == "" {
-				return fmt.Errorf("auth username is missing, required for SNMPv3")
-			}
-		default:
-			return fmt.Errorf("security level must be one of authPriv, authNoPriv or noAuthNoPriv")
-		}
-	}
-	return nil
+	return unmarshal((*plain)(c))
 }
 
 // ConfigureSNMP sets the various version and auth settings.
-func (c WalkParams) ConfigureSNMP(g *gosnmp.GoSNMP) {
+func (c Auth) ConfigureSNMP(g *gosnmp.GoSNMP) {
 	switch c.Version {
 	case 1:
 		g.Version = gosnmp.Version1
@@ -132,16 +107,16 @@ func (c WalkParams) ConfigureSNMP(g *gosnmp.GoSNMP) {
 	case 3:
 		g.Version = gosnmp.Version3
 	}
-	g.Community = string(c.Auth.Community)
-	g.ContextName = c.Auth.ContextName
+	g.Community = string(c.Community)
+	g.ContextName = c.ContextName
 
 	// v3 security settings.
 	g.SecurityModel = gosnmp.UserSecurityModel
 	usm := &gosnmp.UsmSecurityParameters{
-		UserName: c.Auth.Username,
+		UserName: c.Username,
 	}
 	auth, priv := false, false
-	switch c.Auth.SecurityLevel {
+	switch c.SecurityLevel {
 	case "noAuthNoPriv":
 		g.MsgFlags = gosnmp.NoAuthNoPriv
 	case "authNoPriv":
@@ -153,8 +128,8 @@ func (c WalkParams) ConfigureSNMP(g *gosnmp.GoSNMP) {
 		priv = true
 	}
 	if auth {
-		usm.AuthenticationPassphrase = string(c.Auth.Password)
-		switch c.Auth.AuthProtocol {
+		usm.AuthenticationPassphrase = string(c.Password)
+		switch c.AuthProtocol {
 		case "SHA":
 			usm.AuthenticationProtocol = gosnmp.SHA
 		case "SHA224":
@@ -170,8 +145,8 @@ func (c WalkParams) ConfigureSNMP(g *gosnmp.GoSNMP) {
 		}
 	}
 	if priv {
-		usm.PrivacyPassphrase = string(c.Auth.PrivPassword)
-		switch c.Auth.PrivProtocol {
+		usm.PrivacyPassphrase = string(c.PrivPassword)
+		switch c.PrivProtocol {
 		case "DES":
 			usm.PrivacyProtocol = gosnmp.DES
 		case "AES":
@@ -187,6 +162,21 @@ func (c WalkParams) ConfigureSNMP(g *gosnmp.GoSNMP) {
 		}
 	}
 	g.SecurityParameters = usm
+}
+
+type Filters struct {
+	Static  []StaticFilter  `yaml:"static,omitempty"`
+	Dynamic []DynamicFilter `yaml:"dynamic,omitempty"`
+}
+
+type StaticFilter struct {
+	Targets []string `yaml:"targets,omitempty"`
+	Indices []string `yaml:"indices,omitempty"`
+}
+type DynamicFilter struct {
+	Oid     string   `yaml:"oid"`
+	Targets []string `yaml:"targets,omitempty"`
+	Values  []string `yaml:"values,omitempty"`
 }
 
 type Metric struct {
@@ -243,6 +233,46 @@ type Auth struct {
 	PrivProtocol  string `yaml:"priv_protocol,omitempty"`
 	PrivPassword  Secret `yaml:"priv_password,omitempty"`
 	ContextName   string `yaml:"context_name,omitempty"`
+	Version       int    `yaml:"version,omitempty"`
+}
+
+func (c *Auth) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	*c = DefaultAuth
+	type plain Auth
+	if err := unmarshal((*plain)(c)); err != nil {
+		return err
+	}
+
+	if c.Version < 1 || c.Version > 3 {
+		return fmt.Errorf("SNMP version must be 1, 2 or 3. Got: %d", c.Version)
+	}
+	if c.Version == 3 {
+		switch c.SecurityLevel {
+		case "authPriv":
+			if c.PrivPassword == "" {
+				return fmt.Errorf("priv password is missing, required for SNMPv3 with priv")
+			}
+			if c.PrivProtocol != "DES" && c.PrivProtocol != "AES" && c.PrivProtocol != "AES192" && c.PrivProtocol != "AES192C" && c.PrivProtocol != "AES256" && c.PrivProtocol != "AES256C" {
+				return fmt.Errorf("priv protocol must be DES or AES")
+			}
+			fallthrough
+		case "authNoPriv":
+			if c.Password == "" {
+				return fmt.Errorf("auth password is missing, required for SNMPv3 with auth")
+			}
+			if c.AuthProtocol != "MD5" && c.AuthProtocol != "SHA" && c.AuthProtocol != "SHA224" && c.AuthProtocol != "SHA256" && c.AuthProtocol != "SHA384" && c.AuthProtocol != "SHA512" {
+				return fmt.Errorf("auth protocol must be SHA or MD5")
+			}
+			fallthrough
+		case "noAuthNoPriv":
+			if c.Username == "" {
+				return fmt.Errorf("auth username is missing, required for SNMPv3")
+			}
+		default:
+			return fmt.Errorf("security level must be one of authPriv, authNoPriv or noAuthNoPriv")
+		}
+	}
+	return nil
 }
 
 type RegexpExtract struct {
@@ -253,10 +283,7 @@ type RegexpExtract struct {
 func (c *RegexpExtract) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	*c = DefaultRegexpExtract
 	type plain RegexpExtract
-	if err := unmarshal((*plain)(c)); err != nil {
-		return err
-	}
-	return nil
+	return unmarshal((*plain)(c))
 }
 
 // Regexp encapsulates a regexp.Regexp and makes it YAML marshalable.