+++ main.go
@@ -23,6 +23,7 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/alecthomas/kingpin/v2"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
@@ -33,7 +34,6 @@ import (
 	"github.com/prometheus/common/version"
 	"github.com/prometheus/exporter-toolkit/web"
 	webflag "github.com/prometheus/exporter-toolkit/web/kingpinflag"
-	"gopkg.in/alecthomas/kingpin.v2"
 	yaml "gopkg.in/yaml.v2"
 
 	"github.com/prometheus/snmp_exporter/collector"
@@ -41,10 +41,13 @@ import (
 )
 
 var (
-	configFile    = kingpin.Flag("config.file", "Path to configuration file.").Default("snmp.yml").String()
-	webConfig     = webflag.AddFlags(kingpin.CommandLine)
-	listenAddress = kingpin.Flag("web.listen-address", "Address to listen on for web interface and telemetry.").Default(":9116").String()
-	dryRun        = kingpin.Flag("dry-run", "Only verify configuration is valid and exit.").Default("false").Bool()
+	configFile  = kingpin.Flag("config.file", "Path to configuration file.").Default("snmp.yml").String()
+	dryRun      = kingpin.Flag("dry-run", "Only verify configuration is valid and exit.").Default("false").Bool()
+	metricsPath = kingpin.Flag(
+		"web.telemetry-path",
+		"Path under which to expose metrics.",
+	).Default("/metrics").String()
+	toolkitFlags = webflag.AddFlags(kingpin.CommandLine, ":9116")
 
 	// Metrics about the SNMP exporter itself.
 	snmpDuration = promauto.NewSummaryVec(
@@ -52,7 +55,7 @@ var (
 			Name: "snmp_collection_duration_seconds",
 			Help: "Duration of collections by the SNMP exporter",
 		},
-		[]string{"module"},
+		[]string{"auth", "module"},
 	)
 	snmpRequestErrors = promauto.NewCounter(
 		prometheus.CounterOpts{
@@ -66,46 +69,68 @@ var (
 	reloadCh chan chan error
 )
 
+const (
+	proberPath = "/snmp"
+	configPath = "/config"
+)
+
 func handler(w http.ResponseWriter, r *http.Request, logger log.Logger) {
 	query := r.URL.Query()
 
 	target := query.Get("target")
 	if len(query["target"]) != 1 || target == "" {
-		http.Error(w, "'target' parameter must be specified once", 400)
+		http.Error(w, "'target' parameter must be specified once", http.StatusBadRequest)
 		snmpRequestErrors.Inc()
 		return
 	}
 
+	authName := query.Get("auth")
+	if len(query["auth"]) > 1 {
+		http.Error(w, "'auth' parameter must only be specified once", http.StatusBadRequest)
+		snmpRequestErrors.Inc()
+		return
+	}
+	if authName == "" {
+		authName = "public_v2"
+	}
+
 	moduleName := query.Get("module")
 	if len(query["module"]) > 1 {
-		http.Error(w, "'module' parameter must only be specified once", 400)
+		http.Error(w, "'module' parameter must only be specified once", http.StatusBadRequest)
 		snmpRequestErrors.Inc()
 		return
 	}
 	if moduleName == "" {
 		moduleName = "if_mib"
 	}
+
 	sc.RLock()
-	module, ok := (*(sc.C))[moduleName]
+	auth, authOk := sc.C.Auths[authName]
+	module, moduleOk := sc.C.Modules[moduleName]
 	sc.RUnlock()
-	if !ok {
-		http.Error(w, fmt.Sprintf("Unknown module '%s'", moduleName), 400)
+	if !authOk {
+		http.Error(w, fmt.Sprintf("Unknown auth '%s'", authName), http.StatusBadRequest)
+		snmpRequestErrors.Inc()
+		return
+	}
+	if !moduleOk {
+		http.Error(w, fmt.Sprintf("Unknown module '%s'", moduleName), http.StatusBadRequest)
 		snmpRequestErrors.Inc()
 		return
 	}
 
-	logger = log.With(logger, "module", moduleName, "target", target)
+	logger = log.With(logger, "auth", authName, "module", moduleName, "target", target)
 	level.Debug(logger).Log("msg", "Starting scrape")
 
 	start := time.Now()
 	registry := prometheus.NewRegistry()
-	c := collector.New(r.Context(), target, module, logger)
+	c := collector.New(r.Context(), target, auth, module, logger, registry)
 	registry.MustRegister(c)
 	// Delegate http serving to Prometheus client library, which will call collector.Collect.
 	h := promhttp.HandlerFor(registry, promhttp.HandlerOpts{})
 	h.ServeHTTP(w, r)
 	duration := time.Since(start).Seconds()
-	snmpDuration.WithLabelValues(moduleName).Observe(duration)
+	snmpDuration.WithLabelValues(authName, moduleName).Observe(duration)
 	level.Debug(logger).Log("msg", "Finished scrape", "duration_seconds", duration)
 }
 
@@ -118,7 +143,7 @@ func updateConfiguration(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, fmt.Sprintf("failed to reload config: %s", err), http.StatusInternalServerError)
 		}
 	default:
-		http.Error(w, "POST method expected", 400)
+		http.Error(w, "POST method expected", http.StatusBadRequest)
 	}
 }
 
@@ -134,6 +159,12 @@ func (sc *SafeConfig) ReloadConfig(configFile string) (err error) {
 	}
 	sc.Lock()
 	sc.C = conf
+	// Initialize metrics.
+	for auth := range sc.C.Auths {
+		for module := range sc.C.Modules {
+			snmpDuration.WithLabelValues(auth, module)
+		}
+	}
 	sc.Unlock()
 	return nil
 }
@@ -152,10 +183,10 @@ func main() {
 	prometheus.MustRegister(version.NewCollector("snmp_exporter"))
 
 	// Bail early if the config is bad.
-	var err error
-	sc.C, err = config.LoadFile(*configFile)
+	err := sc.ReloadConfig(*configFile)
 	if err != nil {
 		level.Error(logger).Log("msg", "Error parsing config file", "err", err)
+		level.Error(logger).Log("msg", "Possible old config file, see https://github.com/prometheus/snmp_exporter/blob/main/auth-split-migration.md")
 		os.Exit(1)
 	}
 
@@ -163,11 +194,6 @@ func main() {
 	if *dryRun {
 		level.Info(logger).Log("msg", "Configuration parsed successfully")
 		return
-	}
-
-	// Initialize metrics.
-	for module := range *sc.C {
-		snmpDuration.WithLabelValues(module)
 	}
 
 	hup := make(chan os.Signal, 1)
@@ -194,57 +220,77 @@ func main() {
 		}
 	}()
 
-	http.Handle("/metrics", promhttp.Handler()) // Normal metrics endpoint for SNMP exporter itself.
+	http.Handle(*metricsPath, promhttp.Handler()) // Normal metrics endpoint for SNMP exporter itself.
 	// Endpoint to do SNMP scrapes.
-	http.HandleFunc("/snmp", func(w http.ResponseWriter, r *http.Request) {
+	http.HandleFunc(proberPath, func(w http.ResponseWriter, r *http.Request) {
 		handler(w, r, logger)
 	})
 	http.HandleFunc("/-/reload", updateConfiguration) // Endpoint to reload configuration.
 
-	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte(`<html>
-            <head>
-            <title>SNMP Exporter</title>
-            <style>
-            label{
-            display:inline-block;
-            width:75px;
-            }
-            form label {
-            margin: 10px;
-            }
-            form input {
-            margin: 10px;
-            }
-            </style>
-            </head>
-            <body>
-            <h1>SNMP Exporter</h1>
-            <form action="/snmp">
-            <label>Target:</label> <input type="text" name="target" placeholder="X.X.X.X" value="1.2.3.4"><br>
-            <label>Module:</label> <input type="text" name="module" placeholder="module" value="if_mib"><br>
-            <input type="submit" value="Submit">
-            </form>
-						<p><a href="/config">Config</a></p>
-            </body>
-            </html>`))
-	})
+	if *metricsPath != "/" && *metricsPath != "" {
+		landingConfig := web.LandingConfig{
+			Name:        "SNMP Exporter",
+			Description: "Prometheus Exporter for SNMP targets",
+			Version:     version.Info(),
+			Form: web.LandingForm{
+				Action: proberPath,
+				Inputs: []web.LandingFormInput{
+					{
+						Label:       "Target",
+						Type:        "text",
+						Name:        "target",
+						Placeholder: "X.X.X.X/[::X]",
+						Value:       "::1",
+					},
+					{
+						Label:       "Auth",
+						Type:        "text",
+						Name:        "auth",
+						Placeholder: "auth",
+						Value:       "public_v2",
+					},
+					{
+						Label:       "Module",
+						Type:        "text",
+						Name:        "module",
+						Placeholder: "module",
+						Value:       "if_mib",
+					},
+				},
+			},
+			Links: []web.LandingLinks{
+				{
+					Address: configPath,
+					Text:    "Config",
+				},
+				{
+					Address: *metricsPath,
+					Text:    "Metrics",
+				},
+			},
+		}
+		landingPage, err := web.NewLandingPage(landingConfig)
+		if err != nil {
+			level.Error(logger).Log("err", err)
+			os.Exit(1)
+		}
+		http.Handle("/", landingPage)
+	}
 
-	http.HandleFunc("/config", func(w http.ResponseWriter, r *http.Request) {
+	http.HandleFunc(configPath, func(w http.ResponseWriter, r *http.Request) {
 		sc.RLock()
 		c, err := yaml.Marshal(sc.C)
 		sc.RUnlock()
 		if err != nil {
 			level.Error(logger).Log("msg", "Error marshaling configuration", "err", err)
-			http.Error(w, err.Error(), 500)
+			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return
 		}
 		w.Write(c)
 	})
 
-	level.Info(logger).Log("msg", "Listening on address", "address", *listenAddress)
-	srv := &http.Server{Addr: *listenAddress}
-	if err := web.ListenAndServe(srv, *webConfig, logger); err != nil {
+	srv := &http.Server{}
+	if err := web.ListenAndServe(srv, toolkitFlags, logger); err != nil {
 		level.Error(logger).Log("msg", "Error starting HTTP server", "err", err)
 		os.Exit(1)
 	}